Final DB review items: cache qty BigDecimal, pagination caps, SEQUENCE batching (ARCHFIX §1.10/§4.4/§4.5/S8)
§1.10: InventoryBalance.onHandQty + BinStock.quantity/reservedQuantity Integer->BigDecimal(18,3);
V11 widens cache tables idempotently. BinStockResponse stays Integer for frontend compat.

§4.5: PosLayaway.search() capped to 200 most-recent via Pageable. AuditLog sub-endpoints
capped at 500 via Pageable. Main audit log endpoint was already paginated.

S8: Confirmed no cross-tenant leak -- database-per-tenant isolation means findAll() within a
tenant is intentional cross-branch for reports/admin. Branch-scoped lists already go through
branchAccessService.filterBranchScopedByBranch().

§4.4: JournalLine + SalesInvoiceItem switched IDENTITY -> SEQUENCE (allocationSize=50);
V12 creates seq_journal_lines + seq_sales_invoice_items above current max id.
Hibernate JDBC batch props: batch_size=50, order_inserts=true, order_updates=true.

+++ billbull-backend/src/main/java/com/billbull/backend/purchase/invoice/PurchaseInvoiceService.java
@@ -548,7 +548,7 @@ public class PurchaseInvoiceService {
 
             int currentQty = binStockRepository.findByProductId(productId)
                     .stream()
-                    .mapToInt(bs -> bs.getQuantity() != null ? bs.getQuantity() : 0)
+                    .mapToInt(bs -> bs.getQuantity() != null ? bs.getQuantity().intValue() : 0)
                     .sum();
 
             if (!isAgainstGrn && product.isBatch()) {
